Camera manager
Made a script that checks which player the camera is attached to, so now on instance the camera checks whether it's on player1 or player2 then assigns itself a tag so that i can now mess around with the viewport settings. Setting up general architecture so that i can better manipulate game features later on.