respawn logic works
respawn logic works, just need to make sure the death count is increased properly and then make sure that the player does not come back after the 3rd death. So ill relay this to a game manager to make sure that it checks whether or not both players are alive or not.