Perfect Respawn
The players respawn and the timer works perfectly fine. Next i need to fix the hp bar and turn it to a slider, and then Ui for the zones, then I also need to make the general game manager that checks whether both layers are alive or not and whether someone has won or not.