Added VFX and Fixed code
Added vfx for the traps and fixed my code for the burning trap. The freeze code kind of works, it stuns the player, just need to do the same for thundra, then the healing works but goes over cap. Then I needd to cap the trap usage better. THe VFX and Post processing work, post proc is for taking damage and the vfx are for the traps.